fixed bugs:
