+++ Feedbridge/FeedbridgeStandard.swift
@@ -7,6 +7,8 @@
 //
 // swiftlint:disable type_body_length
 // swiftlint:disable file_length
+//
+// Reaoning for swiftlint disable rules: https://github.com/orgs/CS342/discussions/181
 
 import FirebaseAuth
 @preconcurrency import FirebaseFirestore
