+++ Feedbridge/Models/FeedEntry.swift
@@ -12,20 +12,17 @@
 import Foundation
 
 // Represents method of feeding
-// periphery:ignore
 enum FeedType: String, Codable {
     case directBreastfeeding
     case bottle
 }
 
 // Represents the type of milk used
-// periphery:ignore
 enum MilkType: String, Codable {
     case breastmilk
     case formula
 }
 
-// periphery:ignore
 /// Stores feeding-related data
 struct FeedEntry: Identifiable, Codable, Sendable {
     /// Use UUID to generate a unique identifier for Firebase
