+++ Feedbridge/Models/StoolEntry.swift
@@ -12,7 +12,6 @@
 import Foundation
 
 // Represents stool volume classifications
-// periphery:ignore
 enum StoolVolume: String, Codable {
     case light
     case medium
