+++ Feedbridge/Models/WeightEntry.swift
@@ -11,7 +11,6 @@
 @preconcurrency import FirebaseFirestore
 import Foundation
 
-// periphery:ignore
 /// Stores weight measurements (accepts grams, kilograms, or pounds and ounces)
 struct WeightEntry: Identifiable, Codable, Sendable {
     @DocumentID var id: String?
