+++ Feedbridge/Resources/Localizable.xcstrings
@@ -179,16 +179,6 @@
         }
       }
     },
-    "Contacts" : {
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Contacts"
-          }
-        }
-      }
-    },
     "Continue" : {
 
     },
@@ -305,6 +295,7 @@
 
     },
     "LELAND_STANFORD_BIO" : {
+      "extractionState" : "stale",
       "localizations" : {
         "en" : {
           "stringUnit" : {
